Refactor : GroupSearchBar 로직을 SideMenu 내부로 통합하여 구조 단순화
